Refactor import paths for ExpenseTemplate to use the new directory structure
