+++ src/features/Records/EditExpense.tsx
@@ -8,7 +8,7 @@ import { resetEditRecordLS } from '@/shared/utils/records.utils'
 import { DASHBOARD_ROUTE } from '@/shared/constants/Global.constants'
 import { Button } from 'flowbite-react'
 import { RiArrowLeftLine } from '@remixicon/react'
-import { ExpenseTemplate } from './ExpenseTemplate'
+import { ExpenseTemplate } from './ExpenseTemplate/ExpenseTemplate'
 import { useEffect, useState } from 'react'
 import { SelectedAccountLS } from '@/shared/types/global.types'
 import { BankMovement } from '@/shared/types/records.types'
