+++ src/features/Records/ExpenseTemplate/ExpenseTemplate.tsx
@@ -24,17 +24,17 @@ import { ErrorMessage } from "@/shared/ui/atoms/ErrorMessage"
 import { LinkButton } from "@/shared/ui/atoms/LinkButton"
 import { CREATE_EXPENSE_INCOME_ERROR } from "@/shared/constants/records.constants"
 import { CATEGORY_FETCH_ERROR, CATEGORY_REQUIRED, SUBCATEGORY_REQUIRED } from "@/shared/constants/categories.constants"
-import { TransactionCategorizerDropdown } from "../Categories/TransactionCategorizerDropdown"
-import { ManageTagsModal } from "./ManageTagsModal"
+import { TransactionCategorizerDropdown } from "../../Categories/TransactionCategorizerDropdown"
+import { ManageTagsModal } from "../ManageTagsModal"
 import { useManageTags } from "@/shared/hooks/useManageTags"
 import { useIndebtedPeople } from "@/shared/hooks/useIndebtedPeople"
-import { FurtherDetailsAccordeon } from "./FurtherDetailsAccordeon"
+import { FurtherDetailsAccordeon } from "../FurtherDetailsAccordeon"
 import { useMediaQuery } from "@/shared/hooks/useMediaQuery"
-import { PersonalDebtManager } from "../IndebtedPeople/PersonalDebtManager"
+import { PersonalDebtManager } from "../../IndebtedPeople/PersonalDebtManager"
 import { CREDIT_ACCOUNT_TYPE } from "@/shared/types/accounts.types"
 import { Budget, SelectBudget } from "@/shared/types/budgets.types"
 import { useHandleBudgets } from "@/shared/hooks/useHandleBudgets"
-import { SelectBudgetDropdown } from "../Budgets/SelectBudget"
+import { SelectBudgetDropdown } from "../../Budgets/SelectBudget"
 import { BUDGETS_FETCH_ERROR } from "@/shared/constants/budgets.constants"
 
 interface ExpenseTemplateProps {
